Alterando precisao decimal

+++ desafio1.sql
@@ -5,7 +5,7 @@ USE SpotifyClone;
 CREATE TABLE planos(
 plano_id INT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 plano_tipo VARCHAR(25) NOT NULL,
-plano_valor DOUBLE(4, 2) NOT NULL
+plano_valor DECIMAL(4, 2) NOT NULL
 )ENGINE = InnoDB;
 
 CREATE  TABLE artistas(
